@@ -1,6 +1,6 @@
 {
 	"name": "skynovel",
-	"version": "3.12.12",
+	"version": "3.13.0",
 	"publisher": "famibee",
 	"displayName": "SKYNovel",
 	"description": "It supports development of SKYNovel for VSCode, VSCodeでSKYNovelの開発をサポートする拡張機能",
@@ -363,6 +363,10 @@
 			{
 				"command": "skynovel.edPrjJson",
 				"title": "SKYNovel GUIで編集する"
+			},
+			{
+				"command": "skynovel.TempWizard",
+				"title": "テンプレートから始める"
 			}
 		],
 		"viewsContainers": {
@@ -624,6 +628,16 @@
 				]
 			}
 		],
+		"viewsWelcome": [
+			{
+				"view": "skynovel-dev",
+				"contents": "開発環境を調査中です..."
+			},
+			{
+				"view": "skynovel-ws",
+				"contents": "フォルダが開かれていません。\n[テンプレートから始める](command:skynovel.TempWizard)"
+			}
+		],
 		"customEditors": [
 			{
 				"viewType": "SKYNovel.score",
@@ -643,12 +657,14 @@
 		"onView:skynovel-dev"
 	],
 	"dependencies": {
+		"adm-zip": "^0.5.5",
 		"archiver": "^5.3.0",
 		"await-notify": "^1.0.1",
 		"crc-32": "^1.2.0",
 		"crypto-js": "^4.1.1",
 		"fs-extra": "^10.0.0",
 		"image-size": "^1.0.0",
+		"node-fetch": "^2.6.1",
 		"npm-check-updates": "^11.8.3",
 		"png2icons": "^2.0.1",
 		"socket.io": "^4.1.3",
@@ -657,29 +673,31 @@
 		"vscode-debugprotocol": "^1.48.0"
 	},
 	"devDependencies": {
+		"@types/adm-zip": "^0.4.34",
 		"@types/archiver": "^5.3.0",
 		"@types/crypto-js": "^4.0.2",
 		"@types/fs-extra": "^9.0.12",
 		"@types/glob": "^7.1.4",
 		"@types/mocha": "^9.0.0",
-		"@types/node": "^16.6.1",
+		"@types/node": "^16.6.2",
+		"@types/node-fetch": "^2.5.12",
 		"@types/power-assert": "^1.5.4",
 		"@types/uuid": "^8.3.1",
 		"@types/vscode": "^1.59.0",
 		"@typescript-eslint/eslint-plugin": "^4.29.2",
 		"@typescript-eslint/parser": "^4.29.2",
-		"electron": "^13.2.0",
+		"electron": "^13.2.1",
 		"eslint": "^7.32.0",
 		"file-loader": "^6.2.0",
 		"glob": "^7.1.7",
 		"mocha": "^9.0.3",
 		"power-assert": "^1.6.1",
 		"ts-loader": "^9.2.5",
-		"ts-node": "^10.2.0",
+		"ts-node": "^10.2.1",
 		"typescript": "^4.3.5",
 		"vscode-debugadapter-testsupport": "^1.48.0",
 		"vscode-test": "^1.6.1",
-		"webpack": "^5.50.0",
+		"webpack": "^5.51.0",
 		"webpack-cli": "^4.8.0"
 	},
 	"optionalDependencies": {
